feat: make catalogue utils object-based instead of file-based
