+++ .changeset/object-catalogue-deploy.md
@@ -1,0 +1,5 @@
+---
+"jazz-tools": major
+---
+
+BREAKING CHANGE: make `jazz-tools/dev` and `jazz-tools/testing` catalogue deploy helpers accept schema, permissions, and migrations JavaScript objects directly instead of project file paths.
