+++ examples/chat-react/walkthrough/global-setup.ts
@@ -2,8 +2,9 @@
  * Playwright global setup for walkthrough screenshot capture.
  * Starts a local Jazz server and pushes the schema before the web server launches.
  */
-import { join } from "node:path";
 import { type LocalJazzServerHandle, startLocalJazzServer, deploy } from "jazz-tools/testing";
+import permissions from "../permissions.js";
+import { app } from "../schema.js";
 import { WALKTHROUGH_PORT, WALKTHROUGH_APP_ID } from "./walkthrough-constants.js";
 
 const ADMIN_SECRET = "walkthrough-admin-secret";
@@ -23,7 +24,8 @@ export default async function globalSetup() {
     serverUrl: serverHandle.url,
     appId: WALKTHROUGH_APP_ID,
     adminSecret: ADMIN_SECRET,
-    schemaDir: join(import.meta.dirname, ".."),
+    schema: app,
+    permissions,
   });
 
   return async () => {
