+++ examples/docs/todo-server-ts/tests/integration.test.ts
@@ -10,6 +10,8 @@ import { tmpdir } from "node:os";
 import { mkdtempSync } from "node:fs";
 import { join } from "node:path";
 import { deploy, startLocalJazzServer, type LocalJazzServerHandle } from "jazz-tools/testing";
+import permissions from "../permissions.js";
+import { app } from "../schema.js";
 import {
   createServer,
   startServer,
@@ -30,7 +32,8 @@ describe("Todo Server Integration", () => {
       serverUrl: upstream.url,
       appId: upstream.appId,
       adminSecret: upstream.adminSecret,
-      schemaDir: join(import.meta.dirname, ".."),
+      schema: app,
+      permissions,
     });
 
     // Create server with temp persistent storage plus an ephemeral upstream server.
@@ -200,7 +203,8 @@ describe("Todo Server Integration", () => {
         serverUrl: coldStartUpstream.url,
         appId: coldStartUpstream.appId,
         adminSecret: coldStartUpstream.adminSecret,
-        schemaDir: join(import.meta.dirname, ".."),
+        schema: app,
+        permissions,
       });
 
       // Use a shared data path so both server instances see the same Fjall file
@@ -280,7 +284,8 @@ describe("Todo Server Integration", () => {
         serverUrl: sseUpstream.url,
         appId: sseUpstream.appId,
         adminSecret: sseUpstream.adminSecret,
-        schemaDir: join(import.meta.dirname, ".."),
+        schema: app,
+        permissions,
       });
       const sseServer = await startServer(
         await createServer({
