+++ examples/wequencer/e2e/global-setup.ts
@@ -2,9 +2,10 @@
  * Playwright global setup — spawns a local jazz-tools server for sync tests.
  */
 
-import { join } from "node:path";
 import type { FullConfig } from "@playwright/test";
 import { startLocalJazzServer, deploy } from "jazz-tools/testing";
+import permissions from "../permissions.js";
+import { app } from "../schema.js";
 
 const SERVER_PORT = 19878;
 const APP_ID = "00000000-0000-0000-0000-000000000099";
@@ -19,7 +20,8 @@ async function globalSetup(_config: FullConfig): Promise<() => Promise<void>> {
     serverUrl: server.url,
     appId: server.appId,
     adminSecret: server.adminSecret,
-    schemaDir: join(import.meta.dirname, ".."),
+    schema: app,
+    permissions,
   });
 
   return () => server.stop();
