+++ examples/world-tour/tests/browser/global-setup.ts
@@ -1,6 +1,7 @@
-import { join } from "node:path";
 import { deploy, startLocalJazzServer } from "jazz-tools/testing";
 import type { TestProject } from "vitest/node";
+import permissions from "../../permissions.js";
+import { app } from "../../schema.js";
 import { ADMIN_SECRET, APP_ID } from "./test-constants.js";
 
 export async function setup(project: TestProject): Promise<() => Promise<void>> {
@@ -14,7 +15,8 @@ export async function setup(project: TestProject): Promise<() => Promise<void>> 
     serverUrl: handle.url,
     appId: handle.appId,
     adminSecret: ADMIN_SECRET,
-    schemaDir: join(import.meta.dirname, "../.."),
+    schema: app,
+    permissions,
   });
 
   project.provide("worldTourJazzServerUrl", handle.url);
