+++ packages/jazz-tools/src/dev/next.test.ts
@@ -3,6 +3,7 @@ import { access, readFile, writeFile } from "node:fs/promises";
 import { join } from "node:path";
 import { createTempRootTracker, getAvailablePort, todoSchema } from "./test-helpers.js";
 import * as devServer from "./dev-server.js";
+import * as catalogueProject from "./catalogue-project.js";
 import * as schemaWatcher from "./schema-watcher.js";
 import { __resetJazzNextPluginForTests, withJazz, type NextConfigLike } from "./next.js";
 
@@ -199,7 +200,7 @@ describe("withJazz", () => {
     await writeFile(join(schemaDir, "schema.ts"), todoSchema());
 
     const deploy = vi
-      .spyOn(devServer, "deploy")
+      .spyOn(catalogueProject, "deploy")
       .mockRejectedValueOnce(new Error("schema push failed"));
 
     const wrapped = withJazz(
@@ -312,7 +313,7 @@ describe("withJazz", () => {
       stop: vi.fn().mockResolvedValue(undefined),
     });
     const deploy = vi
-      .spyOn(devServer, "deploy")
+      .spyOn(catalogueProject, "deploy")
       .mockResolvedValue(deployed("1111111111111111aaaaaaaaaaaaaaaaaaaaaaaa"));
     let capturedOnPush: ((hash: string) => void) | undefined;
     vi.spyOn(schemaWatcher, "watchSchema").mockImplementation((opts) => {
@@ -359,7 +360,7 @@ describe("withJazz", () => {
       backendSecret: "test-backend-secret",
       stop: vi.fn().mockResolvedValue(undefined),
     });
-    vi.spyOn(devServer, "deploy").mockResolvedValue(deployed("abc"));
+    vi.spyOn(catalogueProject, "deploy").mockResolvedValue(deployed("abc"));
     vi.spyOn(schemaWatcher, "watchSchema").mockImplementation(() => ({ close: vi.fn() }));
 
     const appRoot = await tempRoots.create("jazz-next-alias-");
