+++ packages/jazz-tools/tests/browser/history-conflict.test.ts
@@ -13,11 +13,7 @@ import { describe, it, expect, afterEach, beforeEach } from "vitest";
 import type { Db, TableProxy } from "../../src/runtime/db.js";
 import type { WasmSchema } from "../../src/drivers/types.js";
 import { generateAuthSecret } from "../../src/runtime/auth-secret-store.js";
-import {
-  fetchPermissionsHead,
-  publishStoredPermissions,
-  publishStoredSchema,
-} from "../../src/runtime/schema-fetch.js";
+import { deploy } from "../../src/dev/catalogue.js";
 import {
   getJazzServerInfo,
   unblockJazzServerNetwork,
@@ -78,16 +74,11 @@ describe("History & Conflict Management", () => {
     testingServer = await getJazzServerInfo(uniqueDbName("history-conflict-app"));
     const { appId, serverUrl, adminSecret } = testingServer;
     await unblockJazzServerNetwork(serverUrl);
-    const { hash: schemaHash } = await publishStoredSchema(serverUrl, {
+    await deploy({
       appId,
+      serverUrl,
       adminSecret,
       schema,
-    });
-    const { head } = await fetchPermissionsHead(serverUrl, { appId, adminSecret });
-    await publishStoredPermissions(serverUrl, {
-      appId,
-      adminSecret,
-      schemaHash,
       permissions: {
         todos: {
           select: { using: { type: "True" } },
@@ -99,7 +90,6 @@ describe("History & Conflict Management", () => {
           delete: { using: { type: "True" } },
         },
       },
-      expectedParentBundleObjectId: head?.bundleObjectId ?? null,
     });
   });
